chore: 🚚 renamed stats table to stats
