+++ src/FilamentLogViewerPlugin.php
@@ -7,19 +7,17 @@ use Boquizo\FilamentLogViewer\Entities\Log;
 use Boquizo\FilamentLogViewer\Entities\LogCollection;
 use Boquizo\FilamentLogViewer\Pages\ListLogs;
 use Boquizo\FilamentLogViewer\Pages\ViewLog;
-use Boquizo\FilamentLogViewer\Tables\StatsTable;
 use Boquizo\FilamentLogViewer\UseCases\DeleteLogUseCase;
 use Boquizo\FilamentLogViewer\UseCases\DownloadLogUseCase;
 use Boquizo\FilamentLogViewer\UseCases\DownloadZipUseCase;
 use Boquizo\FilamentLogViewer\UseCases\ExtractLogByDateUseCase;
+use Boquizo\FilamentLogViewer\Utils\Stats;
 use Closure;
 use Filament\Contracts\Plugin;
 use Filament\FilamentManager;
 use Filament\Panel;
 use Filament\Support\Concerns\EvaluatesClosures;
 use Filament\Support\Icons\Heroicon;
-use Illuminate\Support\Facades\Session;
-use RuntimeException;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
 class FilamentLogViewerPlugin implements Plugin
@@ -154,9 +152,9 @@ class FilamentLogViewerPlugin implements Plugin
             ?? __('filament-log-viewer::log.navigation.label');
     }
 
-    public function getViewerStatsTable(): StatsTable
+    public function getViewerStats(): Stats
     {
-        return StatsTable::make((new LogCollection())->stats());
+        return Stats::make((new LogCollection())->stats());
     }
 
     public function getLogsTableFiltered(string $date): array
@@ -170,7 +168,7 @@ class FilamentLogViewerPlugin implements Plugin
     public function getLogsTableRecords(): array
     {
         $rows = $this
-            ->getViewerStatsTable()
+            ->getViewerStats()
             ->rows;
 
         return array_values($rows) ?? [];
